Application service classes moved to specific namespace
