+++ src/Core/Service/ServiceContainerFactoryInterface.php
@@ -8,9 +8,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Core;
+namespace Strider2038\ImgCache\Core\Service;
 
 use Psr\Container\ContainerInterface;
+use Strider2038\ImgCache\Core\ApplicationParameters;
 
 /**
  * @author Igor Lazarev <strider2038@rambler.ru>
