+++ src/Core/Service/ServiceRunnerInterface.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Core;
+namespace Strider2038\ImgCache\Core\Service;
 
 use Psr\Container\ContainerInterface;
